feat: use Laravel Package Tools to load components
